% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlNotify.R
\name{dlNotify}
\alias{dlNotify}
\title{Download and extract tables from europhyt}
\usage{
dlNotify(month, year)
}
\arguments{
\item{month}{Month (integer) of the data to extract}

\item{year}{Year (integer) of the data to extract}
}
\description{
\code{dlNotify} Downloads monthly interceptions data from europhyt
\url{http://ec.europa.eu/food/plant/plant_health_biosecurity/europhyt_en},
and then uses the \code{\link[tabulizer]{extract_tables}} function from the
\code{tabulizer} package to extract the tables.
}
\examples{
oct2016 <- dlNotify(10, 2016)

}

